Fixed missing last byte in jls_core_fsr.
For 1-bit FSR signals, the last byte was incorrectly omitted if the
number of carry bits in the first byte and the number of remaining
bits in the last byte exceed 8 bits when the fetch spanned more than
one FSR data chunk.  In this case, the retrieval must continue for
an extra byte.